requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
digi-xbee==1.4.1
idna==3.6
pillow==10.2.0
pyserial==3.5
requests==2.31.0
six==1.16.0
urllib3==2.2.1
wxPython==4.2.1
